Updated 'Real' to 'f64'

+++ src/base/defaults.rs
@@ -4,7 +4,7 @@
 /// Define the current type in use.
 /// 
 /// Use primitive: `Real` is used in `as` casts.
-pub type Real = f32;
+pub type Real = f64;
 
 /// ULPS level in `float_cmp` for the whole crate. 
-pub(crate) const ULPS: i32 = 2;
+pub(crate) const ULPS: i64 = 2;
